Add support for retreiving FFC metadata
We are using an older version of the Boson SDK because the current
version doesn't compile easily for Raspberry Pi. This older version of
the SDK is lacking APIs for retrieving the frame count of the last FFC
so this has been backported.

When -t is supplied on the command line the FFC details are now
emitted alongside the timing data.